pyproject.toml
[project]
name = "aeolus"
version = "1.0.0"
description = "Generalized WebSocket server with Redis integration for real-time communication"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TutorCruncher", email = "support@tutorcruncher.com"}
]
keywords = ["websocket", "socketio", "redis", "realtime", "chat"]
dependencies = [
    "aiohttp>=3.13.0",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.14.1",
    "redis>=6.4.0",
    "uvicorn>=0.37.0",
]

[project.urls]
Homepage = "https://github.com/tutorcruncher/aeolus"
Repository = "https://github.com/tutorcruncher/aeolus"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [".venv", "__pycache__", "*.pyc"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
]
ignore = [
    "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["aeolus"]
